Luau: mlua 0.12, a hardened sim VM, no FMA; build profiles
- mlua 0.10 -> 0.12.1 (Luau 663 -> 736).
- Sim VM: only deterministic libraries (math, string, table, bit32, utf8,
  buffer); collectgarbage, gcinfo, loadstring, getfenv/setfenv, newproxy and
  os removed; std libraries and globals read-only while rim stays open to
  plugins; each script's environment is a Luau safe env (fast imports,
  fastcalls, fast iteration); compiler level 2; a 256 MB memory limit; a
  counted step budget per call (never wall-clock: it must stop every peer at
  the same point). Script-heavy work: 22.9 -> 13.1 ms.
- UI VM: compiler level 2, safe envs, and a 250 ms deadline per component or
  handler. No memory limit: measured, it made rebuilds 45% slower.
- Luau is built with -ffp-contract=off: clang fused multiply-adds on arm64
  (41 in libluauvm, including the % operator), which could desync a Mac and a
  PC. scripts/check-no-fma.sh checks the build.
- Profiles: line tables only for our crates, no debug info for
  dependencies, profiling and dist profiles (fat LTO, one codegen unit),
  cargo dist / cargo prof aliases.
- Tests: endless loops and runaway memory are stopped, escape hatches are
  gone, libraries are read-only; an endless loop in a UI component becomes an
  error box. examples/luau_bench.rs measures script cost.

+++ crates/rim_sim/src/data.rs
@@ -64,7 +64,7 @@ pub fn from_lua(v: &mlua::Value, path: &str, depth: u32) -> Result<Option<Data>,
     Ok(Some(match v {
         Value::Nil => return Ok(None),
         Value::Boolean(b) => Data::Bool(*b),
-        Value::Integer(i) => Data::Int(*i as i64),
+        Value::Integer(i) => Data::Int(*i),
         Value::Number(n) => {
             if n.fract() == 0.0 && n.abs() < 9.0e15 {
                 Data::Int(*n as i64)
@@ -78,7 +78,7 @@ pub fn from_lua(v: &mlua::Value, path: &str, depth: u32) -> Result<Option<Data>,
             for pair in t.clone().pairs::<Value, Value>() {
                 let (k, v) = pair.map_err(|e| format!("{path}: {e}"))?;
                 let key = match &k {
-                    Value::Integer(i) => Key::Int(*i as i64),
+                    Value::Integer(i) => Key::Int(*i),
                     Value::Number(n) if n.fract() == 0.0 => Key::Int(*n as i64),
                     Value::String(s) => Key::Str(s.to_string_lossy().to_string()),
                     other => {
